backend_ecommerce/requirements.txt
django
djangorestframework
cloudinary
python-decouple
dj-database-url
psycopg2-binary
gunicorn
whitenoise